requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
backcall==0.2.0
bitsandbytes==0.35.4
cachetools==5.3.0
charset-normalizer==3.1.0
clip==0.2.0
cmake==3.26.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
decord==0.6.0
diffusers==0.15.1
docopt==0.6.2
einops==0.6.0
executing==1.2.0
fairscale==0.4.13
filelock==3.10.6
fonttools==4.39.3
ftfy==6.1.1
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
huggingface-hub==0.13.4
idna==3.4
image-reward==1.2
imageio==2.27.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
jedi==0.18.2
jinja2==3.1.2
jupyter-client==8.1.0
jupyter-core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lit==16.0.0
markdown==3.4.3
markupsafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
modelcards==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
pipreqs==0.4.13
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pygments==2.14.0
pyparsing==3.0.9
pyre-extensions==0.0.23
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.0
six==1.16.0
stack-data==0.6.2
supervision==0.6.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
timm==0.6.13
tokenizers==0.13.2
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
triton==2.0.0
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
werkzeug==2.2.3
widgetsnbextension==4.0.7
xformers==0.0.16
yapf==0.33.0
yarg==0.1.9
zipp==3.15.0
scipy==1.10.1
scikit-image==0.21.0
prettytable==3.6.0
moviepy==1.0.3
